requirements.txt
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cycler==0.10.0
demjson==2.2.4
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.0
pandas==1.0.5
patsy==0.5.1
Pillow==7.2.0
pycparser==2.20
pymongo==3.10.1
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
six==1.15.0
statsmodels==0.12.0
threadpoolctl==2.1.0
urllib3==1.25.9
uWSGI==2.0.19.1
Werkzeug==1.0.1
